Refactor dataset-checking code
