--- dsgrid/registry/dimension_association_checker.py
@@ -1,6 +1,0 @@
-class DimensionAssociationChecker:
-    def __init__(self, associations, dimension_mgr, dimension_mapping_mgr, mapping_references):
-        pass
-
-    def check(self):
-        pass
