+++ dsgrid/registry/project_registry_manager.py
@@ -1,22 +1,31 @@
 """Manages the registry for dimension projects"""
-from dsgrid.utils.timing import track_timing
-import getpass
-import logging
 
+import getpass
+import io
+import itertools
+import logging
+from contextlib import redirect_stdout
+from typing import List
+
+from dsgrid.dimension.base_models import DimensionType
+from dsgrid.exceptions import DSGInvalidDataset, DSGValueNotRegistered, DSGDuplicateValueRegistered
 from dsgrid import timer_stats_collector
 from dsgrid.common import REGISTRY_FILENAME
-from dsgrid.config.dimension_mapping_base import DimensionMappingReferenceListModel
+from dsgrid.config.dataset_schema_handler_factory import make_dataset_schema_handler
 from dsgrid.config.dataset_config import DatasetConfig
+from dsgrid.config.dimension_mapping_base import (
+    DimensionMappingReferenceModel,
+    DimensionMappingReferenceListModel,
+)
 from dsgrid.config.project_config import ProjectConfig
-from dsgrid.config.dataset_dimension_checker import check_dataset_dimensions_against_project
-from dsgrid.dataset import Dataset
-from dsgrid.exceptions import DSGValueNotRegistered, DSGDuplicateValueRegistered
 from dsgrid.registry.common import (
     make_initial_config_registration,
     ConfigKey,
     DatasetRegistryStatus,
     ProjectRegistryStatus,
 )
+from dsgrid.utils.spark import create_dataframe_from_dimension_ids
+from dsgrid.utils.timing import track_timing, timer_stats_collector, Timer
 from .common import VersionUpdateType
 from .dataset_registry_manager import DatasetRegistryManager
 from .dimension_registry_manager import DimensionRegistryManager
@@ -210,12 +219,10 @@ class ProjectRegistryManager(RegistryManagerBase):
                 mapping_references.append(ref)
 
         project_config.add_dataset_dimension_mappings(dataset_config, mapping_references)
-        check_dataset_dimensions_against_project(
+        self._check_dataset_base_to_project_base_mappings(
             project_config,
             dataset_config,
             mapping_references,
-            self._dimension_mgr,
-            self._dimension_mapping_mgr,
         )
 
         if self.dry_run_mode:
@@ -247,27 +254,71 @@ class ProjectRegistryManager(RegistryManagerBase):
         )
 
     @track_timing(timer_stats_collector)
-    def check_dataset_base_to_project_base_mappings(
+    def _check_dataset_base_to_project_base_mappings(
         self,
         project_config: ProjectConfig,
         dataset_config: DatasetConfig,
-        dataset: Dataset,
-        mapping_references,
+        mapping_references: List[DimensionMappingReferenceModel],
     ):
-        """Check that the dataset contains all mappings of base dimensions.
+        """Check that a dataset has all project-required dimension records."""
+        handler = make_dataset_schema_handler(
+            dataset_config, self._dimension_mgr, self._dimension_mapping_mgr
+        )
+        pivot_dimension = handler.get_pivot_dimension_type()
+        exclude_dims = set([DimensionType.TIME, DimensionType.DATA_SOURCE, pivot_dimension])
+        dimension_pairs = set()
+        types = [x for x in DimensionType if x not in exclude_dims]
+        for type1, type2 in itertools.product(types, types):
+            if type1 != type2:
+                dimension_pairs.add(tuple(sorted((type1, type2))))
 
-        Parameters
-        ----------
-        project_config : ProjectConfig
-        dataset_config : DatasetConfig
-        dataset : Dataset
+        data_source = dataset_config.model.data_source
+        associations = project_config.dimension_associations
+        dim_table = handler.make_dimension_table(mapping_references)
+        for type1, type2 in dimension_pairs:
+            records = associations.get_associations_by_data_source(data_source, type1, type2)
+            if records is None:
+                records = self._get_project_dimensions_table(project_config, type1, type2)
+            columns = (type1.value, type2.value)
+            with Timer(timer_stats_collector, "evaluate dimension record counts"):
+                record_count = records.count()
+                count = records.select(*columns).intersect(dim_table.select(*columns)).count()
+            if count != record_count:
+                table = records.select(*columns).exceptAll(dim_table.select(*columns))
+                dataset_id = dataset_config.model.dataset_id
+                with io.StringIO() as buf, redirect_stdout(buf):
+                    table.show(n=table.count())
+                    logger.error(
+                        "Dataset %s is missing records for %s:\n%s",
+                        dataset_id,
+                        (type1, type2),
+                        buf.getvalue(),
+                    )
+                raise DSGInvalidDataset(
+                    f"Dataset {dataset_id} is missing records for {(type1, type2)}"
+                )
+        self._check_pivot_dimension_columns(
+            project_config, handler, pivot_dimension, mapping_references
+        )
 
-        Raises
-        ------
-        DSGInvalidDimensionMapping
-            Raised if a requirement is violated.
+    @staticmethod
+    def _get_project_dimensions_table(project_config, type1, type2):
+        pdim1 = project_config.get_base_dimension(type1)
+        pdim1_ids = pdim1.get_unique_ids()
+        pdim2 = project_config.get_base_dimension(type2)
+        pdim2_ids = pdim2.get_unique_ids()
+        records = itertools.product(pdim1_ids, pdim2_ids)
+        return create_dataframe_from_dimension_ids(records, type1, type2)
 
-        """
+    @staticmethod
+    def _check_pivot_dimension_columns(project_config, handler, dim_type, mapping_refs):
+        d_dim_ids = set(
+            handler.get_pivot_dimension_columns_mapped_to_project(mapping_refs).values()
+        )
+        p_dim_ids = project_config.get_base_dimension(dim_type).get_unique_ids()
+        diff = d_dim_ids.difference(p_dim_ids)
+        if diff:
+            raise DSGInvalidDataset(f"load data pivoted columns have invalid values: {diff}")
 
     def update_from_file(
         self, config_file, config_id, submitter, update_type, log_message, version
